requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
firebase-admin==6.2.0
langchain==0.0.350
langchain-openai==0.0.2
openai>=1.6.1,<2.0.0
PyPDF2==3.0.1
pydantic==2.5.0 